feat(stella-pipeline): pre-plan research stage — wiring and protocol (WIP)
