+++ crates/stella-pipeline/src/pipeline/tests/usage.rs
@@ -111,7 +111,7 @@ async fn run_triage_only(
     let result = pipeline
         .triage("inspect the repository", budget, &mut total)
         .await;
-    (result, total, drain(&mut rx))
+    (result.map(|(assessment, _research)| assessment), total, drain(&mut rx))
 }
 
 #[tokio::test]
